refactor: split cli.py into cli/ subpackage
